Add `make install` support for `linux` platform
At the moment, `qmake` generates a Makefile with an empty instruction for
install: due to the fact that `INSTALLS` and `TARGET.{path,files}` were
empty. Now Makefile can perform `make install` / `make uninstall` and
install/delete the executable file, desktop file and icons.

This commit implements this feature only for `linux` platform, but it
can be extended to other unix platforms if desired.

According to the commit history, icons are not used directly anywhere,
not used by flatpack, so moving should not break anything, but it was
necessary to install icons in the correct location on the `linux`
platform.